Compilation Cleanup and Reorganization
Projects can now be built with Debug and Release versions on x86
platform. All binaries now output to the same Bin folder at the root of
the solution directory for easy debugging and release to the installer.
Removed more compilation errors and organized dependencies. Working on a
readme that describes how to install the mod as well as compile and
deploy the source code properly. Will upload tomorrow.